refactor(webview): extract virtual list hook
